% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_sim}
\alias{check_sim}
\title{check_sim}
\usage{
check_sim(fixed, results)
}
\arguments{
\item{fixed}{named list of population level parameters}

\item{results}{named list of simulation results}
}
\description{
check_sim
}
