% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{do_sim}
\alias{do_sim}
\title{do_sim}
\usage{
do_sim(k, mu, tau2, navg, nmin, nsim, alpha = 0.05, summary = TRUE)
}
\arguments{
\item{k}{number of studies}

\item{tau2}{heterogeneity}

\item{navg}{average sample size}

\item{nmin}{minimum sample size}

\item{nsim}{number of simulations}

\item{alpha}{the alpha level}

\item{summary}{logical indicating if the simulation should be summarized using the \code{summary_sim} function.}

\item{theta}{population level effect size}
}
\value{
a dataframe
}
\description{
repeat a simulation a certain number of times. The sample size is sampled from a poisson distribution shifted using \code{nmin}.
}
\examples{
do_sim(10, 0.3, 0.1, 30, 10, 100)
}
