% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{make_data}
\alias{make_data}
\title{make_data}
\usage{
make_data(k, nt = NULL, nc = NULL, ...)
}
\arguments{
\item{k}{the number of studies}

\item{nt}{number indicating the number of participants for the treated group.}

\item{nc}{number indicating the number of participants for the control group. If \code{NULL} the value assigned to \code{nt} will be used.}

\item{...}{other named vectors that will be attached to the dataframe}
}
\value{
a dataframe
}
\description{
a wrapper of \code{data.frame()} to create the grid of simulation values
}
\examples{
make_data(10, 30, 30, x = runif(10)) # k, nt, nc and a column x
}
