% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rma_table}
\alias{rma_table}
\title{rma_table}
\usage{
rma_table(res_tidy, notes, caption)
}
\arguments{
\item{res_tidy}{the first element of the \code{rma_tidy()} output}

\item{notes}{the second element of the \code{rma_tidy()} output}

\item{caption}{the table caption}
}
\description{
rma_table
}
