% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{sim_studies}
\alias{sim_studies}
\title{sim_studies}
\usage{
sim_studies(..., data = NULL)
}
\arguments{
\item{...}{arguments passed to \code{sim_study()}}

\item{data}{a dataframe where the result will be attached (optional). Default to \code{NULL}}
}
\value{
a dataframe
}
\description{
apply the \code{sim_study} function using a grid of parameters.
}
\examples{
sim <- data.frame(k = 1:10, nt = 30, nc = 30)
sim_studies(sim$k, sim$nt, sim$nc, data = sim)
sim_stud
}
