% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{sim_T}
\alias{sim_T}
\title{sim_T}
\usage{
sim_T(k, p, taus2, rho)
}
\arguments{
\item{k}{number of studies}

\item{p}{number of outcomes}

\item{taus2}{vector of tau2 values}

\item{rho}{correlation between population level outcomes}
}
\value{
a numeric vector
}
\description{
Simulate multivariate random-effects assuming a compound symmetry or heteroscedastic compound symmetry structure.
}
\examples{
sim_T(10, 3, c(0.1, 0.1, 0.1), 0.7) # compound symmetry
sim_T(10, 3, c(0.1, 0.05, 0.2), 0.7) # heteroscedastic compound symmetry

}
