% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{summary_sim}
\alias{summary_sim}
\title{summary_sim}
\usage{
summary_sim(p, alpha)
}
\arguments{
\item{p}{the vector of p values}

\item{alpha}{the alpha level}
}
\value{
a dataframe
}
\description{
Summarize the result of \code{do_sim()}
}
\examples{
p <- runif(1000, 0, 1) # random p-values vector
summary_sim(p, 0.05)
}
